feat[xnat]: qunex will exit with error code if it detects error
In the xnat wrapper, after the processing is done, we check the `comlog` and find files
with prefix `error. If such files exist, we exit with code 1; otherwise, we return 0.

XNAT_PROJECT is now also passed to the container as an environment variable in addition
to a command-line argument.

trello: https://trello.com/c/RISvi7lz/837-qunex-container-return-error-code-on-xnat